pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "psy-debate-agent"
version = "0.1.0"
description = "Multi-agent psychiatric triage interview using LangGraph and domestic LLM providers."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "langgraph>=0.2.40",
  "openai>=1.55.0",
  "pydantic>=2.9.0",
  "python-dotenv>=1.0.1",
  "tenacity>=9.0.0",
  "typing-extensions>=4.7.0",
  "sqlalchemy>=2.0.0",
  "pymysql>=1.1.0",
]

[project.scripts]
psy-debate = "psy_debate.main:run"

[tool.setuptools.packages.find]
where = ["src"]
